使用LAB加chardom的算法计算出来的结果也是有很大的出入，暂时不用此方法

+++ C&C++/OpenCV_VS2013/ImageProcess/faceColor.h
@@ -3,4 +3,4 @@
 #include <iostream>
 #include <opencv2/opencv.hpp>
 
-int getFaceColorValue(const std::string &strFile, std::vector<std::vector<cv::Point>> contours);
+double getFaceColorValue(const std::string &strFile, std::vector<std::vector<cv::Point>> contours);
